+++ tests/testthat/test-get_ref_info.R
@@ -142,3 +142,67 @@ test_that("get_ref_info works with a vector analysis function", {
     std_result_matrix
   )
 })
+
+test_that("get_ref_info works with a df in the presence of the overall column", {
+  dm <- formatters::DM
+  dm$colspan_trt <- factor(
+    ifelse(dm$ARM == "B: Placebo", " ", "Active Study Agent"),
+    levels = c("Active Study Agent", " ")
+  )
+  colspan_trt_map <- create_colspan_map(
+    dm,
+    non_active_grp = "B: Placebo",
+    non_active_grp_span_lbl = " ",
+    active_grp_span_lbl = "Active Study Agent",
+    colspan_var = "colspan_trt",
+    trt_var = "ARM"
+  )
+
+  # A standard analysis function which uses a reference group.
+  standard_afun <- function(x, .ref_group, .in_ref_col) {
+    diff_means <- if (!missing(.in_ref_col) && isFALSE(.in_ref_col)) {
+      mean(x) - mean(.ref_group)
+    } else {
+      NULL
+    }
+    in_rows(
+      m = rcell(mean(x), label = "Mean"),
+      dm = rcell(diff_means, label = "Difference in Means vs Placebo"),
+      .formats = "xx.xx"
+    )
+  }
+
+  # The custom analysis function which can work with a global reference group.
+  result_afun <- function(x, ref_path, .spl_context, .var) {
+    ref <- get_ref_info(ref_path, .spl_context, .var)
+    standard_afun(x, .ref_group = ref$ref_group, .in_ref_col = ref$in_ref_col)
+  }
+
+  # Define the global reference group.
+  ref_path <- c("colspan_trt", " ", "ARM", "B: Placebo")
+
+  lyt <- basic_table() |>
+    split_cols_by("colspan_trt", split_fun = trim_levels_to_map(colspan_trt_map)) |>
+    split_cols_by("ARM") |>
+    add_overall_col("Total") |>
+    analyze("AGE", afun = result_afun, extra_args = list(ref_path = ref_path))
+
+  result <- expect_silent(build_table(lyt, dm))
+  expect_snapshot(cran = TRUE, result)
+
+  # Compare with non-hierarchical layout.
+  std_lyt <- basic_table(round_type = "sas") |>
+    split_cols_by("ARM", ref_group = "B: Placebo") |>
+    add_overall_col("Total") |>
+    analyze("AGE", afun = standard_afun, extra_args = list(ref_path = ref_path))
+
+  std_result <- expect_silent(build_table(std_lyt, dm))
+  expect_snapshot(cran = TRUE, std_result)
+
+  result_matrix <- matrix_form(result)$strings
+  std_result_matrix <- matrix_form(std_result)$strings
+  expect_identical(
+    result_matrix[-1, c(1, 2, 4, 3, 5)],
+    std_result_matrix
+  )
+})