+++ NEWS.md
@@ -1,4 +1,15 @@
-# junco 0.1.6
+# junco 0.1.6.9000
+
+### Fixed
+- Fixed `get_ref_info()` so that is works in the presence of "overall" column (#332)
+
+### Changed
+- Removed formatters exports #317
+
+### Added
+
+## [0.1.6] - 2026-05-05 (CRAN release)
+
 
 ### Fixed
 
@@ -44,6 +55,7 @@
 
 ### Added
 
+- Added `a_diff_means()`, which is based on the existing `s_diff_means()`.
 - Added multi-comparator functionality (#271)
 - Hotfix: Added several new functions for creating the Vital Sign tables for core and clinpharm:
   `a_summary_diff_mvars()`, `a_summary_j()`, `c_summary_subset_label()`, `filter_df_prior_afun()`,
@@ -118,6 +130,10 @@
 
 - Initial CRAN release
 
+
+## Changelog
+
+[0.1.6]: https://github.com/johnsonandjohnson/junco/releases/tag/v0.1.6-rc
 [0.1.5]: https://github.com/johnsonandjohnson/junco/releases/tag/v-0.1.5
 [0.1.4]: https://github.com/johnsonandjohnson/junco/releases/tag/0.1.4
 [0.1.3]: https://github.com/johnsonandjohnson/junco/releases/tag/v0.1.3